Less strict matching for spyOn().and.*(..)
Specifically, when matching for `.and.returnValue()`, `.and.callThrough()`,
and `.and.callFake()`, we now accept any preceding object, not only the
common use case of `spyOn().and.*`.

This allows for correct transformation of tests which separate the declaration
of the spy object from the declaration of its behaviour, e.g. by declaring
the spy in `beforeEach()` and configuring it within a test case.

Ref: https://github.com/skovhus/jest-codemods/issues/207